WIP changed Api Traits
