+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [ "zotero_data", "zotero_derive", "zotero_api"]
+resolver = "2"
 
 [workspace.package]
 version = "0.1.0"
@@ -17,7 +18,6 @@ license = "MIT"
 chrono = { version = "0.4.23", features = ["serde"] }
 url = "2.2.2"
 derive_builder = "0.11.2"
-hyperx = { version = "1.4.0", features = ["headers"] }
 regex = "1.7.0"
 serde_json = "1.0.85"
 serde = { version = "1.0.144", features = ["derive"] }
