+++ zotero_api/src/api/delete.rs
