--- zotero_api/src/api_error_parser/mod.rs
@@ -1,2 +1,0 @@
-/// Parse Zotero api Error to custom error
-pub trait ParseZoteroApiError {}
